@@ -8,7 +8,14 @@
   "dependencies": {
     "babel-polyfill": "6.16.0",
     "classnames": "2.2.5",
-    "d3": "4.4.0",
+    "d3-array": "1.0.2",
+    "d3-color": "1.0.2",
+    "d3-format": "1.0.2",
+    "d3-scale": "1.0.4",
+    "d3-selection": "1.0.3",
+    "d3-shape": "1.0.4",
+    "d3-time-format": "2.0.3",
+    "d3-zoom": "1.1.1",
     "dagre": "0.7.4",
     "debug": "2.3.3",
     "filesize": "3.3.0",
@@ -38,7 +45,7 @@
   "devDependencies": {
     "autoprefixer": "6.5.3",
     "babel-core": "6.18.2",
-    "babel-eslint": "5.0.0",
+    "babel-eslint": "7.1.1",
     "babel-jest": "17.0.2",
     "babel-loader": "6.2.8",
     "babel-plugin-lodash": "3.2.10",
@@ -46,11 +53,12 @@
     "babel-preset-react": "6.16.0",
     "clean-webpack-plugin": "0.1.14",
     "css-loader": "0.26.1",
-    "eslint": "2.4.0",
-    "eslint-config-airbnb": "6.1.0",
-    "eslint-loader": "1.3.0",
-    "eslint-plugin-jasmine": "1.6.0",
-    "eslint-plugin-react": "4.2.2",
+    "eslint": "3.11.1",
+    "eslint-config-airbnb": "13.0.0",
+    "eslint-loader": "1.6.1",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "2.2.3",
+    "eslint-plugin-react": "6.8.0",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.9.0",
     "html-webpack-plugin": "2.24.1",
