@@ -14,12 +14,13 @@ class NodeContainer extends React.Component {
     const other = omit(this.props, 'dx', 'dy');
 
     return (
-      <Motion style={{
-        x: spring(dx, animConfig),
-        y: spring(dy, animConfig),
-        f: spring(scaleFactor, animConfig)
-      }}>
-        {interpolated => {
+      <Motion
+        style={{
+          x: spring(dx, animConfig),
+          y: spring(dy, animConfig),
+          f: spring(scaleFactor, animConfig)
+        }}>
+        {(interpolated) => {
           const transform = `translate(${round(interpolated.x, layoutPrecision)},`
             + `${round(interpolated.y, layoutPrecision)})`;
           return <Node {...other} transform={transform} scaleFactor={interpolated.f} />;
