@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames';
-import {getMetricValue, getMetricColor, getClipPathDefinition} from '../utils/metric-utils.js';
-import {CANVAS_METRIC_FONT_SIZE} from '../constants/styles.js';
+import { getMetricValue, getMetricColor, getClipPathDefinition } from '../utils/metric-utils';
+import { CANVAS_METRIC_FONT_SIZE } from '../constants/styles';
 
 
 export default function NodeShapeCircle({id, highlighted, size, color, metric}) {
@@ -17,8 +17,12 @@ export default function NodeShapeCircle({id, highlighted, size, color, metric}) 
       {highlighted && <circle r={size * 0.7} className="highlighted" />}
       <circle r={size * 0.5} className="border" stroke={color} />
       <circle r={size * 0.45} className="shadow" />
-      {hasMetric && <circle r={size * 0.45} className="metric-fill" style={metricStyle}
-        clipPath={`url(#${clipId})`} />}
+      {hasMetric && <circle
+        r={size * 0.45}
+        className="metric-fill"
+        style={metricStyle}
+        clipPath={`url(#${clipId})`}
+      />}
       {highlighted && hasMetric ?
         <text style={{fontSize}}>{formattedValue}</text> :
         <circle className="node" r={Math.max(2, (size * 0.125))} />}
