@@ -1,8 +1,8 @@
 import React from 'react';
 import classNames from 'classnames';
 import { line, curveCardinalClosed } from 'd3-shape';
-import { getMetricValue, getMetricColor, getClipPathDefinition } from '../utils/metric-utils.js';
-import { CANVAS_METRIC_FONT_SIZE } from '../constants/styles.js';
+import { getMetricValue, getMetricColor, getClipPathDefinition } from '../utils/metric-utils';
+import { CANVAS_METRIC_FONT_SIZE } from '../constants/styles';
 
 
 const spline = line()
@@ -12,7 +12,7 @@ const spline = line()
 function polygon(r, sides) {
   const a = (Math.PI * 2) / sides;
   const points = [];
-  for (let i = 0; i < sides; i++) {
+  for (let i = 0; i < sides; i += 1) {
     points.push([r * Math.sin(a * i), -r * Math.cos(a * i)]);
   }
   return points;
@@ -30,15 +30,20 @@ export default function NodeShapeHeptagon({id, highlighted, size, color, metric}
   const metricStyle = { fill: getMetricColor(metric) };
   const className = classNames('shape', { metrics: hasMetric });
   const fontSize = size * CANVAS_METRIC_FONT_SIZE;
+  const halfSize = size * 0.5;
 
   return (
     <g className={className}>
-      {hasMetric && getClipPathDefinition(clipId, size, height, -size * 0.5, size * 0.5 - height)}
+      {hasMetric && getClipPathDefinition(clipId, size, height, -halfSize, halfSize - height)}
       {highlighted && <path className="highlighted" {...pathProps(0.7)} />}
       <path className="border" stroke={color} {...pathProps(0.5)} />
       <path className="shadow" {...pathProps(0.45)} />
-      {hasMetric && <path className="metric-fill" clipPath={`url(#${clipId})`}
-        style={metricStyle} {...pathProps(0.45)} />}
+      {hasMetric && <path
+        className="metric-fill"
+        clipPath={`url(#${clipId})`}
+        style={metricStyle}
+        {...pathProps(0.45)}
+      />}
       {highlighted && hasMetric ?
         <text style={{fontSize}}>{formattedValue}</text> :
         <circle className="node" r={Math.max(2, (size * 0.125))} />}
