@@ -1,8 +1,8 @@
 import React from 'react';
 import classNames from 'classnames';
 import { line, curveCardinalClosed } from 'd3-shape';
-import { getMetricValue, getMetricColor, getClipPathDefinition } from '../utils/metric-utils.js';
-import { CANVAS_METRIC_FONT_SIZE } from '../constants/styles.js';
+import { getMetricValue, getMetricColor, getClipPathDefinition } from '../utils/metric-utils';
+import { CANVAS_METRIC_FONT_SIZE } from '../constants/styles';
 
 
 const spline = line()
@@ -48,13 +48,22 @@ export default function NodeShapeHexagon({id, highlighted, size, color, metric})
 
   return (
     <g className={className}>
-      {hasMetric && getClipPathDefinition(clipId,
-        size * (1 + hexCurve * 2), height, -size * hexCurve, (size - height) * shadowSize * 2)}
+      {hasMetric && getClipPathDefinition(
+        clipId,
+        size * (1 + (hexCurve * 2)),
+        height,
+        -(size * hexCurve),
+        (size - height) * (shadowSize * 2)
+      )}
       {highlighted && <path className="highlighted" {...pathProps(0.7)} />}
       <path className="border" stroke={color} {...pathProps(0.5)} />
       <path className="shadow" {...pathProps(shadowSize)} />
-      {hasMetric && <path className="metric-fill" style={metricStyle}
-        clipPath={`url(#${clipId})`} {...pathProps(shadowSize)} />}
+      {hasMetric && <path
+        className="metric-fill"
+        style={metricStyle}
+        clipPath={`url(#${clipId})`}
+        {...pathProps(shadowSize)}
+      />}
       {highlighted && hasMetric ?
         <text style={{fontSize}}>
           {formattedValue}
