@@ -9,10 +9,12 @@ class Details extends React.Component {
     // render all details as cards, later cards go on top
     return (
       <div className="details">
-        {details.toIndexedSeq().map((obj, index) => <DetailsCard key={obj.id}
-          index={index} cardCount={details.size}
-          nodeControlStatus={controlStatus.get(obj.id)} {...obj} />
-        )}
+        {details.toIndexedSeq().map((obj, index) => (
+          <DetailsCard
+            key={obj.id} index={index} cardCount={details.size}
+            nodeControlStatus={controlStatus.get(obj.id)} {...obj}
+          />
+        ))}
       </div>
     );
   }
