@@ -4,8 +4,10 @@ import LogMonitor from 'redux-devtools-log-monitor';
 import DockMonitor from 'redux-devtools-dock-monitor';
 
 export default createDevTools(
-  <DockMonitor defaultIsVisible={false}
-    toggleVisibilityKey="ctrl-h" changePositionKey="ctrl-w">
+  <DockMonitor
+    defaultIsVisible={false}
+    toggleVisibilityKey="ctrl-h"
+    changePositionKey="ctrl-w">
     <LogMonitor />
   </DockMonitor>
 );
