@@ -4,6 +4,21 @@ import classNames from 'classnames';
 
 import { toggleGridMode } from '../actions/app-actions';
 
+
+const Item = (icons, label, isSelected, onClick) => {
+  const className = classNames('grid-mode-selector-action', {
+    'grid-mode-selector-action-selected': isSelected
+  });
+  return (
+    <div
+      className={className}
+      onClick={onClick} >
+      <span className={icons} style={{fontSize: 12}} />
+      <span>{label}</span>
+    </div>
+  );
+};
+
 class GridModeSelector extends React.Component {
 
   constructor(props, context) {
@@ -21,28 +36,14 @@ class GridModeSelector extends React.Component {
     return this.props.toggleGridMode(false);
   }
 
-  renderItem(icons, label, isSelected, onClick) {
-    const className = classNames('grid-mode-selector-action', {
-      'grid-mode-selector-action-selected': isSelected
-    });
-    return (
-      <div
-        className={className}
-        onClick={onClick} >
-        <span className={icons} style={{fontSize: 12}} />
-        <span>{label}</span>
-      </div>
-    );
-  }
-
   render() {
     const { gridMode } = this.props;
 
     return (
       <div className="grid-mode-selector">
         <div className="grid-mode-selector-wrapper">
-          {this.renderItem('fa fa-share-alt', 'Graph', !gridMode, this.disableGridMode)}
-          {this.renderItem('fa fa-table', 'Table', gridMode, this.enableGridMode)}
+          {Item('fa fa-share-alt', 'Graph', !gridMode, this.disableGridMode)}
+          {Item('fa fa-table', 'Table', gridMode, this.enableGridMode)}
         </div>
       </div>
     );
