@@ -6,25 +6,22 @@ import MatchedText from './matched-text';
 const SHOW_ROW_COUNT = 2;
 const MAX_MATCH_LENGTH = 24;
 
+
+const Match = match => (
+  <div className="matched-results-match" key={match.label}>
+    <div className="matched-results-match-wrapper">
+      <span className="matched-results-match-label">
+        {match.label}:
+      </span>
+      <MatchedText
+        text={match.text} match={match}
+        maxLength={MAX_MATCH_LENGTH}
+        truncate={match.truncate} />
+    </div>
+  </div>
+);
+
 class MatchedResults extends React.Component {
-
-  renderMatch(matches, field) {
-    const match = matches.get(field);
-    const text = match.text;
-
-    return (
-      <div className="matched-results-match" key={match.label}>
-        <div className="matched-results-match-wrapper">
-          <span className="matched-results-match-label">
-            {match.label}:
-          </span>
-          <MatchedText text={text} match={match} maxLength={MAX_MATCH_LENGTH}
-            truncate={match.truncate} />
-        </div>
-      </div>
-    );
-  }
-
   render() {
     const { matches, style } = this.props;
 
@@ -44,7 +41,7 @@ class MatchedResults extends React.Component {
 
     return (
       <div className="matched-results" style={style}>
-        {matches.keySeq().take(SHOW_ROW_COUNT).map(fieldId => this.renderMatch(matches, fieldId))}
+        {matches.keySeq().take(SHOW_ROW_COUNT).map(fieldId => Match(matches.get(fieldId)))}
         {moreFieldMatches && <div className="matched-results-more" title={moreFieldMatchesTitle}>
           {`${moreFieldMatches.size} more matches`}
         </div>}
