@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { connect } from 'react-redux';
 
 import { clickRelative } from '../../actions/app-actions';
@@ -10,18 +9,29 @@ class NodeDetailsRelativesLink extends React.Component {
   constructor(props, context) {
     super(props, context);
     this.handleClick = this.handleClick.bind(this);
+    this.saveNodeRef = this.saveNodeRef.bind(this);
   }
 
   handleClick(ev) {
     ev.preventDefault();
-    this.props.dispatch(clickRelative(this.props.id, this.props.topologyId,
-      this.props.label, ReactDOM.findDOMNode(this).getBoundingClientRect()));
+    this.props.dispatch(clickRelative(
+      this.props.id,
+      this.props.topologyId,
+      this.props.label,
+      this.node.getBoundingClientRect()
+    ));
+  }
+
+  saveNodeRef(ref) {
+    this.node = ref;
   }
 
   render() {
     const title = `View in ${this.props.topologyId}: ${this.props.label}`;
     return (
-      <span className="node-details-relatives-link" title={title} onClick={this.handleClick}>
+      <span
+        className="node-details-relatives-link" title={title}
+        onClick={this.handleClick} ref={this.saveNodeRef}>
         <MatchedText text={this.props.label} match={this.props.match} />
       </span>
     );
