@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { connect } from 'react-redux';
 
 import { clickRelative } from '../../actions/app-actions';
@@ -9,12 +8,21 @@ class NodeDetailsTableNodeLink extends React.Component {
   constructor(props, context) {
     super(props, context);
     this.handleClick = this.handleClick.bind(this);
+    this.saveNodeRef = this.saveNodeRef.bind(this);
   }
 
   handleClick(ev) {
     ev.preventDefault();
-    this.props.dispatch(clickRelative(this.props.nodeId, this.props.topologyId,
-      this.props.label, ReactDOM.findDOMNode(this).getBoundingClientRect()));
+    this.props.dispatch(clickRelative(
+      this.props.nodeId,
+      this.props.topologyId,
+      this.props.label,
+      this.node.getBoundingClientRect()
+    ));
+  }
+
+  saveNodeRef(ref) {
+    this.node = ref;
   }
 
   render() {
@@ -23,8 +31,9 @@ class NodeDetailsTableNodeLink extends React.Component {
 
     if (linkable) {
       return (
-        <span className="node-details-table-node-link" title={title}
-          onClick={this.handleClick}>
+        <span
+          className="node-details-table-node-link" title={title}
+          ref={this.saveNodeRef} onClick={this.handleClick}>
           {label}
         </span>
       );
