@@ -113,13 +113,13 @@ function getNodeValue(node, header) {
 
 
 function getValueForSortedBy(sortedByHeader) {
-  return (node) => maybeToLower(getNodeValue(node, sortedByHeader));
+  return node => maybeToLower(getNodeValue(node, sortedByHeader));
 }
 
 
 function getMetaDataSorters(nodes) {
   // returns an array of sorters that will take a node
-  return get(nodes, [0, 'metadata'], []).map((field, index) => node => {
+  return get(nodes, [0, 'metadata'], []).map((field, index) => (node) => {
     const nodeMetadataField = node.metadata && node.metadata[index];
     if (nodeMetadataField) {
       if (isNumber(nodeMetadataField)) {
@@ -225,7 +225,7 @@ export default class NodeDetailsTable extends React.Component {
       <tr>
         {headers.map((header, i) => {
           const headerClasses = ['node-details-table-header', 'truncate'];
-          const onHeaderClick = ev => {
+          const onHeaderClick = (ev) => {
             this.handleHeaderClick(ev, header.id, sortedBy, sortedDesc);
           };
           // sort by first metric by default
@@ -243,7 +243,8 @@ export default class NodeDetailsTable extends React.Component {
             header.label;
 
           return (
-            <td className={headerClasses.join(' ')} style={style} onClick={onHeaderClick}
+            <td
+              className={headerClasses.join(' ')} style={style} onClick={onHeaderClick}
               title={header.label} key={header.id}>
               {isSortedAsc
                 && <span className="node-details-table-header-sorter fa fa-caret-up" />}
@@ -278,14 +279,15 @@ export default class NodeDetailsTable extends React.Component {
     const className = classNames('node-details-table-wrapper-wrapper', this.props.className);
 
     return (
-      <div className={className}
-        style={this.props.style}>
+      <div className={className} style={this.props.style}>
         <div className="node-details-table-wrapper">
           <table className="node-details-table">
             <thead>
               {this.renderHeaders(sortedBy, sortedDesc)}
             </thead>
-            <tbody style={this.props.tbodyStyle} onMouseEnter={onMouseEnter}
+            <tbody
+              style={this.props.tbodyStyle}
+              onMouseEnter={onMouseEnter}
               onMouseLeave={onMouseLeave}>
               {nodes && nodes.map(node => (
                 <NodeDetailsTableRow
