@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import NodesChart from '../charts/nodes-chart';
 import NodesGrid from '../charts/nodes-grid';
 import NodesError from '../charts/nodes-error';
-import { DelayedShow } from '../utils/delayed-show';
+import DelayedShow from '../utils/delayed-show';
 import { Loading, getNodeType } from './loading';
 import { isTopologyEmpty } from '../utils/topology-utils';
 import { CANVAS_MARGINS } from '../constants/styles';
@@ -12,6 +12,20 @@ import { CANVAS_MARGINS } from '../constants/styles';
 const navbarHeight = 194;
 const marginTop = 0;
 
+
+const EmptyTopologyError = show => (
+  <NodesError faIconClass="fa-circle-thin" hidden={!show}>
+    <div className="heading">Nothing to show. This can have any of these reasons:</div>
+    <ul>
+      <li>We haven&apos;t received any reports from probes recently.
+       Are the probes properly configured?</li>
+      <li>There are nodes, but they&apos;re currently hidden. Check the view options
+       in the bottom-left if they allow for showing hidden nodes.</li>
+      <li>Containers view only: you&apos;re not running Docker,
+       or you don&apos;t have any containers.</li>
+    </ul>
+  </NodesError>
+);
 
 class Nodes extends React.Component {
   constructor(props, context) {
@@ -32,22 +46,6 @@ class Nodes extends React.Component {
     window.removeEventListener('resize', this.handleResize);
   }
 
-  renderEmptyTopologyError(show) {
-    return (
-      <NodesError faIconClass="fa-circle-thin" hidden={!show}>
-        <div className="heading">Nothing to show. This can have any of these reasons:</div>
-        <ul>
-          <li>We haven't received any reports from probes recently.
-           Are the probes properly configured?</li>
-          <li>There are nodes, but they're currently hidden. Check the view options
-           in the bottom-left if they allow for showing hidden nodes.</li>
-          <li>Containers view only: you're not running Docker,
-           or you don't have any containers.</li>
-        </ul>
-      </NodesError>
-    );
-  }
-
   render() {
     const { topologyEmpty, gridMode, topologiesLoaded, nodesLoaded, topologies,
       currentTopology } = this.props;
@@ -60,16 +58,11 @@ class Nodes extends React.Component {
             itemType={getNodeType(currentTopology, topologies)}
             show={topologiesLoaded && !nodesLoaded} />
         </DelayedShow>
-        {this.renderEmptyTopologyError(topologiesLoaded && nodesLoaded && topologyEmpty)}
+        {EmptyTopologyError(topologiesLoaded && nodesLoaded && topologyEmpty)}
 
         {gridMode ?
-          <NodesGrid {...this.state}
-            nodeSize="24"
-            margins={CANVAS_MARGINS}
-          /> :
-         <NodesChart {...this.state}
-           margins={CANVAS_MARGINS}
-           />}
+          <NodesGrid {...this.state} nodeSize="24" margins={CANVAS_MARGINS} /> :
+          <NodesChart {...this.state} margins={CANVAS_MARGINS} />}
       </div>
     );
   }
