@@ -4,6 +4,18 @@ import classnames from 'classnames';
 
 import { clickTopology } from '../actions/app-actions';
 
+
+function basicTopologyInfo(topology, searchMatchCount) {
+  const info = [
+    `Nodes: ${topology.getIn(['stats', 'node_count'])}`,
+    `Connections: ${topology.getIn(['stats', 'edge_count'])}`
+  ];
+  if (searchMatchCount) {
+    info.push(`Search Matches: ${searchMatchCount}`);
+  }
+  return info.join('\n');
+}
+
 class Topologies extends React.Component {
 
   constructor(props, context) {
@@ -21,29 +33,21 @@ class Topologies extends React.Component {
     const topologyId = subTopology.get('id');
     const searchMatches = this.props.searchNodeMatches.get(subTopology.get('id'));
     const searchMatchCount = searchMatches ? searchMatches.size : 0;
-    const title = this.renderTitle(subTopology, searchMatchCount);
+    const title = basicTopologyInfo(subTopology, searchMatchCount);
     const className = classnames('topologies-sub-item', {
       'topologies-sub-item-active': isActive,
       'topologies-sub-item-matched': searchMatchCount
     });
 
     return (
-      <div className={className} title={title} key={topologyId} rel={topologyId}
+      <div
+        className={className} title={title} key={topologyId} rel={topologyId}
         onClick={this.onTopologyClick}>
         <div className="topologies-sub-item-label">
           {subTopology.get('name')}
         </div>
       </div>
     );
-  }
-
-  renderTitle(topology, searchMatchCount) {
-    let title = `Nodes: ${topology.getIn(['stats', 'node_count'])}\n`
-      + `Connections: ${topology.getIn(['stats', 'node_count'])}`;
-    if (searchMatchCount) {
-      title = `${title}\nSearch Matches: ${searchMatchCount}`;
-    }
-    return title;
   }
 
   renderTopology(topology) {
@@ -55,7 +59,7 @@ class Topologies extends React.Component {
       'topologies-item-main-matched': searchMatchCount
     });
     const topologyId = topology.get('id');
-    const title = this.renderTitle(topology, searchMatchCount);
+    const title = basicTopologyInfo(topology, searchMatchCount);
 
     return (
       <div className="topologies-item" key={topologyId}>
