@@ -149,8 +149,13 @@ export default ComposedComponent => class extends React.Component {
       .filter(dateFilter);
 
     const lastValue = samples.length > 0 ? samples[samples.length - 1].value : null;
-    const slidingWindow = {first: movingFirstDate,
-      last: movingLastDate, max, samples, value: lastValue};
+    const slidingWindow = {
+      first: movingFirstDate,
+      last: movingLastDate,
+      value: lastValue,
+      samples,
+      max
+    };
 
     return <ComposedComponent {...this.props} {...slidingWindow} />;
   }
