@@ -1,10 +1,10 @@
-require('../styles/main.less');
-require('../images/favicon.ico');
-
 import 'babel-polyfill';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+
+import '../styles/main.less';
+import '../images/favicon.ico';
 import configureStore from './stores/configureStore';
 
 const store = configureStore();
