@@ -1,4 +1,4 @@
-import _ from 'lodash';
+import { zipObject } from 'lodash';
 import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
 
@@ -104,10 +104,9 @@ function mergeMetrics(node) {
     return node;
   }
   return Object.assign({}, node, {
-    metrics: _(metrics[node.shape])
+    metrics: (metrics[node.shape] || [])
       .map((fn, name) => [name, fn(node)])
       .fromPairs()
-      .value()
   });
 }
 
@@ -121,7 +120,7 @@ function handleAdd(nodes) {
 
 
 function handleUpdated(updatedNodes, prevNodes) {
-  const modifiedNodesIndex = _.zipObject((updatedNodes || []).map(n => [n.id, n]));
+  const modifiedNodesIndex = zipObject((updatedNodes || []).map(n => [n.id, n]));
   return prevNodes.toIndexedSeq().toJS().map(n => (
     Object.assign({}, mergeMetrics(n), modifiedNodesIndex[n.id])
   ));
